Less intrusive diagnostics for template syntax errors.
Since we don't get specific location info from `@glimmer/syntax`
for a syntax error, we're forced to just indicate the whole template
is broken.

Doing this by emitting a diagnostic that spans the full string is
disruptive as you're trying to type out a template, so instead we
emit the diagnostic against the template tag instead.