core/cli: add `--build` and related commands
Introduce new flags in the CLI, along with `--build` itself:

* `--dry`
* `--force`
* `--incremental`
* `--clean`
* `projects`

These options map directly to the `tsc` options, and allow users to
actually *use* the newly-implemented build functionality.

Introduce the check for running a `--build` *before* falling into the
default case, so that `--build` and especially `--build --watch` has
the right behavior.

This is straightforward *except* for the hacky use of a default command
to capture the optional positional arguments allowed by `tsc` with its
`--build` options. That pattern for `yargs` is documented [here][1], so
is considered supported even if a bit odd.

[1]: https://github.com/yargs/yargs/blob/main/docs/advanced.md#default-commands