core/cli: introduce `performBuild` and `performBuildWatch`
These correspond to the `--build` and `--build --watch` commands. Glint
forwards to the normal TS compiler API for these commands, but with a
patched version of the `TS.Program` type which incorporates the
`TransformManager` hooks so that running `--build` (with or without
`--watch`) incorporates transformed templates and can properly emit
diagnostics for them.