core/cli: do not test against `.tsbuildinfo` files
Whether these are generated is even more implementation-specified than
TS features in general (which is saying something). Between TS 4.7 and
TS 4.8, the answer changes. We *hope* it will be stable after that, and
it may be that we can reintroduce them after TS 4.8, but for now, focus
instead on the generation of `.d.ts` files, which is a far more robust
artifact.

When introducing support for `--incremental` and `--watch`, it may make
more sense to do `statSync` checks on the emitted `.d.ts` files to show
they have or have not changed as expected.