core/cli: implement `glint --build --dry`
This should also include the bits required to support `--incremental`
builds, since if TS *knows* that a build needs to be re-run and can
report it in `--dry`, it can also *do* the build with `--incremental`.

As part of this effort, I noticed that the tests were assuming we would
emit `.js` files, but that is actually incorrect. We should *only* emit
`.d.ts` files, including during `--build`.