chore: switch to `rmdir(path, { recursive: true })`
Using `rmdir` with `{ recursive }` is deprecated. Noticed it in console
output while debugging tests, so removing it now. Additionally, for all
the test setups, we need to make sure to do `force: true` as well. The
new invocation is:

    fs.rmSync(path, { recursive: true, force: true })

(This makes sure that it does not blow up when the file does not exist.)