core/cli: add `glint --incremental` support
Include the flag in the CLI and incorporate it for both `--build` and
regular builds.

The test here is intentionally minimal: we exercise a much broader set
of features via the `--build` and especially `--build --dry` tests,
which exercise these capabilities explicitly. Here, we just check that
`--incremental` does the right thing when *not* in a `--build` context.