[reword] core/cli: fix `--incremental` builds
When users pass `--incremental` to a regular check invocation (not a
`--watch` invocation), configure the host and program by using the
variants TS supplies:

- `createIncrementalCompilerHost()`
- `createIncrementalProgram()`

This makes the program *aware* of incremental builds. However, to
actually work correctly, we also need to collect diagnostics *before*
running `program.emit()`, or else TS will not cache the diagnostics in
the `.tsbuildinfo` file it gnerates. This is odd, but it's sort of
documented in an unmerged PR to the TS wiki; see [here][docs], and
incremental builds work correctly once using this strategy.

[docs]: ttps://github.com/microsoft/TypeScript-wiki/blob/ad7afb1b7049be5ac59ba55dce9a647390ee8481/Using-the-Compiler-API.md#a-minimal-incremental-compiler