core/cli: Introduce tests for `--build --watch`
Since `--build` is tested *extensively*, and these tests are quite
expensive, introduce a minimal set of tests:

- Working projects pass and shut down correctly.
- Changes which produce either type errors or parse errors report those
  errors, and when the errors are fixed, the watch returns its base "no
  errors" state.

Confirm that this works for both basic and composite projects, since
users can call `--build --watch` in both cases. For composite projects,
test the following combinations:

- errors in the root
- errors in a project referenced directly in the root
- errors in a project referenced transitively from the root.

For reasons that are not entirely clear, the composite project tests
require the insertion of arbitrary additional wait time between
launching the project and making changes, and between additional
changes, for the output to be captured in our tests. Accordingly, in
several places below, you will see invocations like:

    await pauseFor(1000);

This may be a result of IO buffering managed by the test runner, by TS
trying to be smart about bundling up changes, or some combination of
the two. For the moment, we are landing these as is, unless/until we
get reports of actual problems in real-world code.