Make `Get` helper interop with DT and preview types
In the `intrinsics` module for the `@glint/environment-ember-loose`,
add a declaration for the `ComputedPropertyMarker`. If using the DT
types, this acts as a declaration merge for a module which is then
integrated via the rest of the DT types. If using the stable/preview
types, this ends up being a no-op.