Introduce a test package using stable/preview types
This package intentionally does *not* have any of the runtime config or
dependencies; it only needs `ember-source` and the Glint packages to be
able to verify that type checking works correctly using the types
published as part of the `ember-source` package. An annoying exception:
we include the ESLint deps since we run ESLint from the root. We might
be able to fix that in the future, but for now this unblocks things.

This package also includes a check to make sure that this test package
is *not* accidentally (read: via incorrect dependency hoisting) getting
the DT types.