Add support for `--build --verbose` invocations
This is super handy for debugging when a `--build` is failing for some
reason. The default reporting from `tsc` or `glint` itself isn't always
helpful: it tends to bottom out with a *symptom* like "some type is not
defined" without any way to get to the root cause of *why* that type is
not defined. Having `--verbose` available for a `--build` invocation
surfaces additional information, including which `references` are being
built or not, and why.